lint fix and dep cycle fix
