+++ frontend/workflows/projectSelector/src/types.tsx
@@ -62,6 +62,13 @@ export interface DashState {
   // Contains the names of selected projects, upstreams, and downstreams merged together.
   selected: string[];
 
-  // Contains a map of project names to the full project data. 
+  // Contains a map of project names to the full project data.
   projectData: { [projectName: string]: IClutch.core.project.v1.IProject };
+}
+
+export type DashActionKind = "UPDATE_SELECTED";
+
+export interface DashAction {
+  type: DashActionKind;
+  payload: DashState;
 }
